Move wallet RPC declarations to rpcwallet.h
