+++ src/rpc/misc.cpp
@@ -16,6 +16,7 @@
 #include "utilstrencodings.h"
 #include "validation.h"
 #ifdef ENABLE_WALLET
+#include "wallet/rpcwallet.h"
 #include "wallet/wallet.h"
 #include "wallet/walletdb.h"
 #endif
@@ -385,6 +386,9 @@ UniValue validateaddress(const JSONRPCRequest& request)
     }
     return ret;
 }
+
+// Needed even with !ENABLE_WALLET, to pass (ignored) pointers around
+class CWallet;
 
 /**
  * Used by addmultisigaddress / createmultisig:
