+++ src/rpc/rawtransaction.cpp
@@ -26,6 +26,7 @@
 #include "utilstrencodings.h"
 #include "instantx.h"
 #ifdef ENABLE_WALLET
+#include "wallet/rpcwallet.h"
 #include "wallet/wallet.h"
 #endif
