+++ src/wallet/rpcdump.cpp
@@ -17,6 +17,8 @@
 #include "merkleblock.h"
 #include "core_io.h"
 
+#include "rpcwallet.h"
+
 #include <fstream>
 #include <stdint.h>
