+++ src/wallet/rpcwallet.h
@@ -6,7 +6,20 @@
 #define BITCOIN_WALLET_RPCWALLET_H
 
 class CRPCTable;
+class JSONRPCRequest;
 
 void RegisterWalletRPCCommands(CRPCTable &t);
+
+/**
+ * Figures out what wallet, if any, to use for a JSONRPCRequest.
+ *
+ * @param[in] request JSONRPCRequest that wishes to access a wallet
+ * @return NULL if no wallet should be used, or a pointer to the CWallet
+ */
+CWallet *GetWalletForJSONRPCRequest(const JSONRPCRequest&);
+
+std::string HelpRequiringPassphrase(CWallet *);
+void EnsureWalletIsUnlocked(CWallet *);
+bool EnsureWalletIsAvailable(CWallet *, bool avoidException);
 
 #endif //BITCOIN_WALLET_RPCWALLET_H
